MultiSSH+RedScripts/requirements.txt
asyncssh>=2.14.0
pyyaml>=6.0
click>=8.1.0
rich>=13.0.0
aiofiles>=23.0.0